wip: adding integration tests
